remove assistant integrations CLI and relocate shared utilities
Delete the assistant integrations CLI command (telegram, twilio, voice
subcommands), its barrel re-export, and test file. Move shared CLI
utilities (writeOutput, shouldOutputJson, gatewayGet, etc.) from
integrations/utils.ts to a standalone cli/utils.ts so other CLI modules
that depended on the barrel keep working. Update all consumer imports,
reference help text, retrieval guard allowlist, and documentation.
